Refactor container config
This commit handle the TODO task of breaking the Container
config into smaller sub-configs
